+++ src/backend/projects.ts
@@ -3,7 +3,7 @@ import type { ProjectBody } from "@/types";
 import fs from "fs";
 import path from "path";
 import crypto from "crypto";
-import { dialog } from "electron";
+import { app, dialog } from "electron";
 
 import {
   DEFAULT_WINDOW_TITLE,
@@ -11,10 +11,12 @@ import {
   EXISTING_DATA_MESSAGE,
   EXISTING_DATA_BUTTONS,
   PROJECT_FILE_NAME,
+  PROJECT_EXPORT_DIRECTORY,
   INITIAL_MATCHED_STACKS,
 } from "@/constants";
+import { getAlphabetLetter } from "@/helpers";
 
-import { createPhotoThumbnail } from "@/backend/photos";
+import { createPhotoEditsCopy, createPhotoThumbnail } from "@/backend/photos";
 import { addRecentProject } from "@/backend/recents";
 
 const sendData = (mainWindow: Electron.BrowserWindow, data: ProjectBody) => {
@@ -27,6 +29,9 @@ const sendData = (mainWindow: Electron.BrowserWindow, data: ProjectBody) => {
   });
 };
 
+const homePath = app.getPath("home");
+const desktopPath = path.resolve(homePath, "Desktop");
+
 /**
  * Handles opening, filtering, and processing a project folder.
  */
@@ -34,6 +39,7 @@ const handleOpenDirectoryPrompt = async (mainWindow: Electron.BrowserWindow) => 
   const event = await dialog.showOpenDialog({
     title: "Open Project Folder",
     properties: ["openDirectory"],
+    defaultPath: desktopPath,
   });
 
   if (event.canceled) {
@@ -42,7 +48,7 @@ const handleOpenDirectoryPrompt = async (mainWindow: Electron.BrowserWindow) => 
 
   const [directory] = event.filePaths;
 
-  const files = fs.readdirSync(directory);
+  const files = await fs.promises.readdir(directory);
 
   if (files.includes(PROJECT_FILE_NAME)) {
     const { response } = await dialog.showMessageBox({
@@ -59,7 +65,7 @@ const handleOpenDirectoryPrompt = async (mainWindow: Electron.BrowserWindow) => 
 
     // Pre-existing project
     if (response === 1) {
-      const data = fs.readFileSync(path.join(directory, PROJECT_FILE_NAME), "utf8");
+      const data = await fs.promises.readFile(path.join(directory, PROJECT_FILE_NAME), "utf8");
       return sendData(mainWindow, JSON.parse(data) as ProjectBody);
     }
 
@@ -86,9 +92,10 @@ const handleOpenDirectoryPrompt = async (mainWindow: Electron.BrowserWindow) => 
     return true;
   });
 
-  const thumbnails = await Promise.all(
-    photos.map((photo) => createPhotoThumbnail(photo, directory)),
-  );
+  const [edited, thumbnails] = await Promise.all([
+    Promise.all(photos.map((photo) => createPhotoEditsCopy(photo, directory))),
+    Promise.all(photos.map((photo) => createPhotoThumbnail(photo, directory))),
+  ]);
 
   const now = new Date().toISOString();
 
@@ -108,6 +115,7 @@ const handleOpenDirectoryPrompt = async (mainWindow: Electron.BrowserWindow) => 
     totalPhotos: photos.length,
     photos: photos.map((name, index) => ({
       name,
+      edited: edited[index],
       thumbnail: thumbnails[index],
     })),
     matched: defaultMatches,
@@ -116,7 +124,11 @@ const handleOpenDirectoryPrompt = async (mainWindow: Electron.BrowserWindow) => 
     lastModified: now,
   };
 
-  fs.writeFileSync(path.join(directory, PROJECT_FILE_NAME), JSON.stringify(data, null, 2), "utf8");
+  await fs.promises.writeFile(
+    path.join(directory, PROJECT_FILE_NAME),
+    JSON.stringify(data, null, 2),
+    "utf8",
+  );
 
   return sendData(mainWindow, data);
 };
@@ -129,6 +141,7 @@ const handleOpenFilePrompt = async (mainWindow: Electron.BrowserWindow) => {
     title: "Open Project File",
     properties: ["openFile"],
     filters: [{ name: "Projects", extensions: ["json"] }],
+    defaultPath: desktopPath,
   });
 
   if (event.canceled) {
@@ -140,26 +153,61 @@ const handleOpenFilePrompt = async (mainWindow: Electron.BrowserWindow) => {
 
   const [file] = event.filePaths;
 
-  const data = fs.readFileSync(file, "utf8");
+  const data = await fs.promises.readFile(file, "utf8");
   return sendData(mainWindow, JSON.parse(data) as ProjectBody);
 };
 
 /**
  * Handles opening a recent project file.
  */
-const handleOpenProjectFile = (mainWindow: Electron.BrowserWindow, file: string) => {
+const handleOpenProjectFile = async (mainWindow: Electron.BrowserWindow, file: string) => {
   mainWindow.webContents.send("set-loading", true, "Opening project");
 
-  const data = fs.readFileSync(file, "utf8");
+  const data = await fs.promises.readFile(file, "utf8");
   return sendData(mainWindow, JSON.parse(data) as ProjectBody);
 };
 
 /**
  * Handles saving a project file.
  */
-const handleSaveProject = (data: string) => {
+const handleSaveProject = async (data: string) => {
   const { directory } = JSON.parse(data) as ProjectBody;
-  fs.writeFileSync(path.join(directory, PROJECT_FILE_NAME), data, "utf8");
+  await fs.promises.writeFile(path.join(directory, PROJECT_FILE_NAME), data, "utf8");
+};
+
+/**
+ * Handles exporting matches.
+ */
+const handleExportMatches = async (data: string) => {
+  const project = JSON.parse(data) as ProjectBody;
+
+  const exportsDirectory = path.join(project.directory, PROJECT_EXPORT_DIRECTORY);
+  if (!fs.existsSync(exportsDirectory)) {
+    await fs.promises.mkdir(exportsDirectory);
+  } else {
+    // Empty exports folder
+    for (const file of await fs.promises.readdir(exportsDirectory)) {
+      await fs.promises.unlink(path.join(exportsDirectory, file));
+    }
+  }
+
+  for (const match of project.matched) {
+    const matchID = getAlphabetLetter(match.id);
+
+    for (const photo of match.left.photos) {
+      const exportedName = `${match.left.name.padStart(3, "0").toUpperCase() || matchID}L_${photo.name}`;
+      const originalPath = path.join(project.directory, photo.edited);
+      const exportedPath = path.join(exportsDirectory, exportedName);
+      await fs.promises.copyFile(originalPath, exportedPath);
+    }
+
+    for (const photo of match.right.photos) {
+      const exportedName = `${match.right.name.padStart(3, "0").toUpperCase() || matchID}R_${photo.name}`;
+      const originalPath = path.join(project.directory, photo.edited);
+      const exportedPath = path.join(exportsDirectory, exportedName);
+      await fs.promises.copyFile(originalPath, exportedPath);
+    }
+  }
 };
 
 export {
@@ -167,4 +215,5 @@ export {
   handleOpenFilePrompt,
   handleOpenProjectFile,
   handleSaveProject,
+  handleExportMatches,
 };